pyproject.toml
[tool.poetry]
name = "fin_work"
version = "0.1.0"
description = ""
authors = ["higashi <higashigo@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
pip = "^22.1.2"
install = "^1.3.5"
django-filter = "^22.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
